--- a/bench/Bench.hs
+++ b/bench/Bench.hs
@@ -3,54 +3,44 @@
 {-# LANGUAGE OverloadedStrings #-}
 module Main where
 
-import Control.DeepSeq         (NFData)
 import Criterion.Main
-import Data.ProtocolBuffers
-import GHC.Generics(Generic)
-import Data.Binary.Builder.Sized
-import Data.Binary.Get
-import Data.ByteString(ByteString)
-import qualified Data.ByteString.Lazy as LBS
-
-data Burst = Burst {
-    frames :: Repeated 1 (Value ByteString)
-} deriving (Generic, Eq, Show)
-
-instance NFData Burst
-instance Encode Burst
-instance Decode Burst
-
-burst :: Int -> Burst
-burst = Burst . putField . flip replicate "abcd"
-
-encodeBurst :: Int -> LBS.ByteString
-encodeBurst = enc . burst
-
-enc :: Encode a => a -> LBS.ByteString
-enc = toLazyByteString . encodeMessage
-
-dec :: Decode a => LBS.ByteString -> a
-dec = runGet decodeMessage
-
-decBurst :: LBS.ByteString -> Burst
-decBurst = dec
+import Data
+import Nested
 
 main :: IO ()
 main = defaultMain [
-  bgroup "encoding" [
-     bgroup "burst" [
+  bgroup "burst" [
+     bgroup "encoding" [
         bench "1" $ nf enc (burst 1),
         bench "10" $ nf enc (burst 10),
         bench "100" $ nf enc (burst 100),
         bench "1000" $ nf enc (burst 1000)
-        ]
-     ],
-  bgroup "decoding" [
-     bgroup "burst" [
+        ],
+     bgroup "decoding" [
         bench "1" $ nf decBurst (encodeBurst 1),
         bench "10" $ nf decBurst (encodeBurst 10),
         bench "100" $ nf decBurst (encodeBurst 100),
         bench "1000" $ nf decBurst (encodeBurst 1000)
+        ]
+  ],
+  bgroup "nested" [
+     bgroup "encoding" [
+        bench "a1" $ nf enc (nested 1 1),
+        bench "a10" $ nf enc (nested 1 10),
+        bench "a100" $ nf enc (nested 1 100),
+        bench "a1000" $ nf enc (nested 1 1000),
+        bench "b10" $ nf enc (nested 10 1),
+        bench "b100" $ nf enc (nested 100 1),
+        bench "b1000" $ nf enc (nested 1000 1)
+     ],
+     bgroup "decoding" [
+        bench "a1" $ nf decNested (enc $ nested 1 1),
+        bench "a10" $ nf decNested (enc $ nested 1 10),
+        bench "a100" $ nf decNested (enc $ nested 1 100),
+        bench "a1000" $ nf decNested (enc $ nested 1 1000),
+        bench "b10" $ nf decNested (enc $ nested 10 1),
+        bench "b100" $ nf decNested (enc $ nested 100 1),
+        bench "b1000" $ nf decNested (enc $ nested 1000 1)
         ]
      ]
   ]