--- a/src/Data/ProtocolBuffers/Encode.hs
+++ b/src/Data/ProtocolBuffers/Encode.hs
@@ -16,7 +16,8 @@ import Data.Foldable
 import Data.HashMap.Strict (HashMap)
 import qualified Data.HashMap.Strict as HashMap
 import Data.Proxy
-import Data.Binary.Put
+import Data.Binary.Builder
+import Data.Monoid
 
 import GHC.Generics
 import GHC.TypeLits
@@ -27,44 +28,43 @@ import qualified Data.ByteString.Lazy as LBS
 
 -- |
 -- Encode a Protocol Buffers message.
-encodeMessage :: Encode a => a -> Put
+encodeMessage :: Encode a => a -> Builder
 encodeMessage = encode
 
 -- |
 -- Encode a Protocol Buffers message prefixed with a varint encoded 32-bit integer describing it's length.
-encodeLengthPrefixedMessage :: Encode a => a -> Put
+encodeLengthPrefixedMessage :: Encode a => a -> Builder
 {-# INLINE encodeLengthPrefixedMessage #-}
-encodeLengthPrefixedMessage msg = do
-  let msg' = runPut $ encodeMessage msg
-  putVarUInt $ LBS.length msg'
-  putLazyByteString msg'
+encodeLengthPrefixedMessage msg = (putVarUInt $ LBS.length msg') <> fromLazyByteString msg'
+  where
+    msg' = toLazyByteString $ encodeMessage msg
 
 class Encode (a :: *) where
-  encode :: a -> Put
-  default encode :: (Generic a, GEncode (Rep a)) => a -> Put
+  encode :: a -> Builder
+  default encode :: (Generic a, GEncode (Rep a)) => a -> Builder
   encode = gencode . from
 
 -- | Untyped message encoding
 instance Encode (HashMap Tag [WireField]) where
-  encode = traverse_ step . HashMap.toList where
-    step = uncurry (traverse_ . encodeWire)
+  encode = foldMap step . HashMap.toList where
+    step = uncurry (foldMap . encodeWire)
 
 class GEncode (f :: * -> *) where
-  gencode :: f a -> Put
+  gencode :: f a -> Builder
 
 instance GEncode a => GEncode (M1 i c a) where
   gencode = gencode . unM1
 
 instance (GEncode a, GEncode b) => GEncode (a :*: b) where
-  gencode (x :*: y) = gencode x >> gencode y
+  gencode (x :*: y) = gencode x <> gencode y
 
 instance (GEncode a, GEncode b) => GEncode (a :+: b) where
   gencode (L1 x) = gencode x
   gencode (R1 y) = gencode y
 
 instance (EncodeWire a, KnownNat n, Foldable f) => GEncode (K1 i (Field n (f a))) where
-  gencode = traverse_ (encodeWire tag) . runField . unK1 where
+  gencode = foldMap (encodeWire tag) . runField . unK1 where
     tag = fromIntegral $ natVal (Proxy :: Proxy n)
 
 instance GEncode U1 where
-  gencode _ = return ()
+  gencode _ = empty