@NathanHowell 2015-06-14
I'm just about to head out to Bayhac and I'll review the patch sometime today. I'd also like to see some heap profiles.

Back when we first started work on protobuf, binary was missing necessary functionality that cereal had, I think it was lacking an incremental parser. The last attempt (a year or two ago) at switching didn't show that much of performance difference, but they've been doing a lot more work on binary so I'm not surprised the decodes are magically faster.