@joshbohde 2015-06-14
Using the following code, modified for respective branch:

    {-# LANGUAGE DataKinds #-}
    {-# LANGUAGE DeriveGeneric #-}
    {-# LANGUAGE OverloadedStrings #-}
    module Main where
    
    import Data
    import Nested
    import qualified Data.ByteString as BS
    
    writeBurst :: IO ()
    writeBurst = do
      BS.writeFile "burst.pb" $ enc $ burst 1000000
    
    readBurst :: IO ()
    readBurst = do
      a <- BS.readFile "burst.pb"
      print $ decBurst a
    
    writeNested :: IO ()
    writeNested = do
      BS.writeFile "nested.pb" $ enc $ nested 100 1000
    
    readNested :: IO ()
    readNested = do
      a <- BS.readFile "nested.pb"
      print $ decNested a
    
    main = readNested

* Write Burst
** Master
![master_writeburst](https://cloud.githubusercontent.com/assets/58237/8149589/2fe96d60-128e-11e5-87ef-605a8f5e90ac.png)
** Binary
![binary_writeburst](https://cloud.githubusercontent.com/assets/58237/8149593/2fefda60-128e-11e5-8248-7310af1d004d.png)

* Read Burst
** Master
![master_readburst](https://cloud.githubusercontent.com/assets/58237/8149587/2fe0ea5a-128e-11e5-89fe-3453a21c0486.png)
** Binary
![binary_readburst](https://cloud.githubusercontent.com/assets/58237/8149591/2fee0bb8-128e-11e5-93b8-a960ab983657.png)

* Write Nested
** Master
![master_writenested](https://cloud.githubusercontent.com/assets/58237/8149588/2fe21664-128e-11e5-8631-91b897b042a5.png)
** Binary
![binary_writenested](https://cloud.githubusercontent.com/assets/58237/8149592/2fee32b4-128e-11e5-9a59-7935f840cadd.png)

* Read Nested
** Master
![master_readnested](https://cloud.githubusercontent.com/assets/58237/8149586/2fe025ac-128e-11e5-9650-454e38bf82aa.png)
** Binary
![binary_readnested](https://cloud.githubusercontent.com/assets/58237/8149590/2fedd22e-128e-11e5-8e91-5d37b0bba3d9.png)